update tests, restructure files
