+++ .codeocean/environment.json
@@ -28,6 +28,29 @@
 			],
 			"options": {},
 			"pre_install_options": {}
+		},
+		"vscode": {
+			"packages": [
+				{
+					"name": "REditorSupport.R"
+				},
+				{
+					"name": "continue.continue"
+				},
+				{
+					"name": "ms-python.python"
+				},
+				{
+					"name": "ms-toolsai.jupyter"
+				},
+				{
+					"name": "reageyao.bioSyntax"
+				},
+				{
+					"name": "saoudrizwan.claude-dev"
+				}
+			],
+			"version": "4.95.3"
 		}
 	}
 }
