+++ code/job_dispatch/__init__.py
@@ -1,0 +1,1 @@
+"""Package for analysis job dispatch"""
